requirements.txt
allure-pytest==2.15.2
allure-python-commons==2.15.2
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
colorama==0.4.6
h11==0.16.0
idna==3.11
iniconfig==2.3.0
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==9.0.2
pytweening==1.2.0
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.2
websocket-client==1.9.0
wsproto==1.3.2
